Initial version of akka http based benchmark server
