--- src/main/scala/org/fneves/benchmark/BenchmarkServer.scala
@@ -1,22 +1,0 @@
-package org.fneves.benchmark
-
-import com.twitter.finagle.httpx.{Request, Response}
-import com.twitter.finatra.http.HttpServer
-import com.twitter.finatra.http.filters.CommonFilters
-import com.twitter.finatra.http.routing.HttpRouter
-import com.twitter.finatra.logging.filter.{LoggingMDCFilter, TraceIdMDCFilter}
-import com.twitter.finatra.logging.modules.Slf4jBridgeModule
-
-object BenchmarkerServerMain extends BenchmarkServer
-
-class BenchmarkServer extends HttpServer {
-  override def modules = Seq(Slf4jBridgeModule)
-
-  override def configureHttp(router: HttpRouter) {
-    router
-      .filter[LoggingMDCFilter[Request, Response]]
-      .filter[TraceIdMDCFilter[Request, Response]]
-      .filter[CommonFilters]
-      .add[BenchmarkController]
-  }
-}
